// src/app/model/user.model.ts
// export interface FormData {
//   id?: number; 
//   userName: string;
//   email: string;
//   mobile: string;
//   password: string;
//   role?: string; 
//   }

  export interface PageMaster {
    pageMasterId: number;
    pageName: string;
    pageSlug: string;
    createdBy: string;
    createdByName: string;
    isActive: boolean;
  }